pyproject.toml
[project]
name = "smartguard-vision"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "bcrypt<5",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
